requirements_for_docker.txt
# --------- pytorch --------- #
pytorch-lightning
torchmetrics #==0.11.1

# --------- hydra --------- #
hydra-core #==1.3.1
hydra-colorlog #==1.2.0
hydra-optuna-sweeper #==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

# --------- data augmentation --------- #
albumentations #>=1.3.0
dlib
opencv-python
numpy